Validate the return query-param
The `return` query-param holds the URL where the response is returned.
It is set by the caller and should correspond to one of the SAML
DiscoveryResponse elements.

Since this is controlled by the caller, We must ensure that it is
correct and sanitise it. Ideally, we should compare it against the
Location attribute of the known <DiscoveryResponse> elements.

If the `return` query-param is not a valid URL we throw an error.